backend/requirements.txt
selenium==4.17.2
pytesseract==0.3.10
pillow==10.2.0
opencv-python==4.9.0.80
webdriver-manager
boto3
python-dotenv